sample_apps/sdksampleapps/python/agent_example/pyproject.toml
[project]
name = "pipeshub-sdk-agent-example"
version = "0.1.0"
description = "Sample application demonstrating PipesHub Agents API using the Python SDK (pipeshub-sdk)"
requires-python = ">=3.10"
dependencies = [
    "pipeshub-sdk>=0.1.2",
    "python-dotenv>=1.0.0",
    "requests>=2.28.0",
]

[project.scripts]
agent-example = "main:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
